services/search/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "search-service"
version = "0.1.0"
description = "Search Service for D&D Character Creator"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "search_service", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.0"
uvicorn = {extras = ["standard"], version = "^0.23.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.20"}
alembic = "^1.11.3"
asyncpg = "^0.28.0"
pydantic = {extras = ["email"], version = "^2.3.0"}
pydantic-settings = "^2.0.3"
elasticsearch = {extras = ["async"], version = "^8.10.0"}
redis = {extras = ["hiredis"], version = "^5.0.0"}
prometheus-client = "^0.17.0"
structlog = "^23.1.0"
aio-pika = "^9.3.0"
httpx = "^0.24.1"
nltk = "^3.8.1"
spacy = "^3.7.0"
scikit-learn = "^1.3.0"
sentence-transformers = "^2.2.2"
transformers = "^4.34.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
ruff = "^0.0.286"
mypy = "^1.5.1"

[tool.poetry.scripts]
start = "search_service.main:start"
dev = "search_service.main:dev"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
warn_no_return = true
warn_return_any = true
strict_equality = true
disallow_any_generics = false

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
asyncio_mode = "auto"
addopts = """
    --strict-markers
    --tb=short
    --cov=src
    --cov-report=term-missing
    --cov-report=xml
    --cov-report=html
    --no-cov-on-fail
"""
